IceTouch installer: simplications to uninstall command

+++ distribution/src/mac/IceTouch/uninstall.sh
@@ -1,11 +1,14 @@
 #!/bin/sh
 
+#
+# This script uninstall all IceTouch-1.1.1 distributions installed in the computer.
+#
+
 showHelp ()
 {
-    echo "ZeroC, IceTouch uninstaller script"
+    echo "ZeroC, IceTouch-1.1.1 uninstaller script"
     echo "usage: "
-    echo "  \"sudo $0 --mode all\" uninstall all IceTouch components"
-    echo "  \"sudo $0\" unistal\" uninstall all IceTouch components except IceTouch command line SDK"
+    echo "  \"sudo $0\" -- uninstall all IceTouch-1.1.1 distributions installed in the computer."
 }
 
 #
@@ -16,13 +19,11 @@ uninstallPackage ()
     PACKAGE=$1
 
     if [[ ! -f "/var/db/receipts/$PACKAGE.bom" ]]; then
-        echo "Package '"$PACKAGE"' not installed"
-        return 1
+        return 0
     fi
 
     if [[ ! -f "/var/db/receipts/$PACKAGE.plist" ]]; then
-        echo "Package '"$PACKAGE"' not installed"
-        return 1
+        return 0
     fi
     
     VOLUME=""
@@ -49,7 +50,6 @@ uninstallPackage ()
         VERSION_MAJOR=${VERSION:0:1}
         VERSION_MINOR=${VERSION:2:1}
         VERSION_MM=$VERSION_MAJOR"."$VERSION_MINOR
-
 
         XCODE_DEV_PACKAGE="NO"
         if [[ "$PACKAGE" == "com.zeroc.icetouch-xcode41-developer.pkg" ]]; then
@@ -80,29 +80,35 @@ uninstallPackage ()
         fi
 
         if [[ "$PACKAGE" == "com.zeroc.icetouch-command-line-developer.pkg" ]]; then
-            rm -rf "/opt/IceTouch-$VERSION_MM"
+            if [[ "/opt/IceTouch-$VERSION_MM" ]]; then
+                rm -rf "/opt/IceTouch-$VERSION_MM"
+            fi
         fi
 
         if [[ "$PACKAGE" == "com.zeroc.icetouch-xcode32-plugin.pkg" ]]; then
-            rm -rf "/Library/Application Support/Developer/3.2/Xcode/Plug-ins/slice2objcplugin.pbplugin"
+            if [[ -d "/Library/Application Support/Developer/3.2/Xcode/Plug-ins/slice2objcplugin.pbplugin" ]]; then
+                rm -rf "/Library/Application Support/Developer/3.2/Xcode/Plug-ins/slice2objcplugin.pbplugin"
+            fi
         fi
 
         if [[ "$PACKAGE" == "com.zeroc.icetouch-xcode40-plugin.pkg" ]]; then
-            rm -rf "/Library/Application Support/Developer/4.0/Xcode/Plug-ins/slice2objcplugin.pbplugin"
+            if [[ -d "/Library/Application Support/Developer/4.0/Xcode/Plug-ins/slice2objcplugin.pbplugin" ]]; then
+                rm -rf "/Library/Application Support/Developer/4.0/Xcode/Plug-ins/slice2objcplugin.pbplugin"
+            fi
         fi
 
         if [[ "$PACKAGE" == "com.zeroc.icetouch-xcode41-plugin.pkg" ]]; then
-            rm -rf "/Library/Application Support/Developer/4.1/Xcode/Plug-ins/slice2objcplugin.pbplugin"
+            if [[ -d "/Library/Application Support/Developer/4.1/Xcode/Plug-ins/slice2objcplugin.pbplugin" ]]; then
+                rm -rf "/Library/Application Support/Developer/4.1/Xcode/Plug-ins/slice2objcplugin.pbplugin"
+            fi
         fi
 
-        echo "Package: '"$PACKAGE"' uninstalled ok"
         rm "/var/db/receipts/$PACKAGE.bom"
         rm "/var/db/receipts/$PACKAGE.plist"
+
         return 0
     }
 }
-
-mode=""
 
 #
 # Parse command line arguments.
@@ -110,34 +116,24 @@ mode=""
 while true; do
     case "$1" in
       -h|--help|-\?) showHelp; exit 0;;
-      --mode) if [ $# -gt 1 ]; then
-            mode=$2
-            shift 2
-            if [[ $mode != "all" ]]; then
-                echo "valid modes are 'all'"
-                showHelp
-                exit 1
-            fi
-          else 
-            echo "-f requires an argument" 1>&2
-            exit 1
-          fi ;;
-      --) shift; break;;
       -*) echo "invalid option: $1" 1>&2; showHelp; exit 1;;
       *)  break;;
     esac
 done
 
+#
+# Display a error if not running as root.
+#
 if [ "$(id -u)" != "0" ]; then
     echo "ERROR: Must be run with root permissions -- prefix command with 'sudo'" 1>&2
     showHelp
     exit 1
 fi
 
-
-#
-# Uninstall IceTouch packages
-#
-if [[ $mode == "all" ]]; then
-    uninstallPackage "com.zeroc.icetouch-command-line-developer.pkg"
-fi
+uninstallPackage "com.zeroc.icetouch-command-line-developer.pkg"
+uninstallPackage "com.zeroc.icetouch-xcode32-developer.pkg"
+uninstallPackage "com.zeroc.icetouch-xcode40-developer.pkg"
+uninstallPackage "com.zeroc.icetouch-xcode41-developer.pkg"
+uninstallPackage "com.zeroc.icetouch-xcode32-plugin.pkg"
+uninstallPackage "com.zeroc.icetouch-xcode40-plugin.pkg"
+uninstallPackage "com.zeroc.icetouch-xcode41-plugin.pkg"
